requirements.txt
torchvision==0.8.1
matplotlib==3.2.0
yacs==0.1.8
scipy==1.5.4
torch==1.7.0
numpy==1.18.0
advertorch==0.2.3
Pillow==8.3.1
scikit_learn==0.24.2
tensorboardX==2.4
